% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{FebuxostatVsAllopurinolCVD}
\alias{FebuxostatVsAllopurinolCVD}
\alias{FebuxostatVsAllopurinolCVD-package}
\title{FebuxostatVsAllopurinolCVD}
\description{
FebuxostatVsAllopurinolCVD
}
